pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "insta-news"
version = "0.1.0"
description = "UniMelb Club Oracle - Club events to your inbox"
authors = [{name = "Aaron"}]
requires-python = ">=3.10"

dependencies = [
    "fastapi[standard]>=0.109.0",
    "instaloader>=4.14.0",
    "google-genai>=0.1.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.27.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
# teaching note: repo root finds `src.*` packages; `src/web` finds `email_system` (how main.py imports).
pythonpath = [".", "src/web"]
addopts = "-q --tb=short"

# teaching note:
# Declares packages under the repo root so `pip install -e .` adds them to the
# environment and imports like `import src.web.main` work from any cwd.
[tool.setuptools.packages.find]
where = ["."]
